Move prepopulation to splash screen
Run it as the last request, after the blocking requests
